requirements.txt
matplotlib==3.3.2
mlflow==1.11.0
optuna==2.1.0
pandas==1.1.2
Pillow==7.2.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
